fix(test): deduplicate TSIG test helpers (#8009)
